+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - Add ``IsTrue``.
 - Add ``IsFalse``.
 - Make ``Except`` **not** call *value* if it's callable. (**breaking change**)
+- Make ``Except`` return ``True`` if all assertions pass.
 
 
 v0.0.1 (2015-05-07)